feat: add pidlock, only run updates for active sessions
